Remove case_sensitive find option
Applying case insensitivity during the find/matching process is
inefficient, especially if the collection being searched is going to
live beyond a single search. Consumers are much better off mapping the
values to lowercase on their own. Eventually, this method will take an
iterator, making this functionality possible again, using an adapter
to convert the values to lowercase.