Specify AsStr trait bound on type, not reference
Previously, the AsStr trait bound was specified directly for type T, but
we're expecting the iterator to yield references, not types. As it was,
AsStr would have to be implemented for references. This new signature
better expresses what we're after: an iterator that yields references to
types that implement AsStr.